requirements.txt
argparse-1.4.0
dataclasses==0.6
firebase-admin==5.2.0
pandas==1.3.5
python-decouple==3.3
